integrate(6822): record main module extraction evidence
