+++ specs/6822-split-main-module-extraction-contract.md
@@ -20,12 +20,12 @@ Split `crates/kamn-node/tests/main_module_extraction_contract.rs` into a thin ro
 - Extraction contract markers drift from the real root layout
 
 ## Acceptance criteria
-- [ ] `crates/kamn-node/tests/main_module_extraction_contract.rs` is reduced to a thin root shell at or below 180 LOC
-- [ ] Root shell wires bounded sibling modules for the current main-module extraction concerns and shared support
-- [ ] All extracted files touched by the split remain at or below 200 LOC
-- [ ] `cargo test -p kamn-node --test main_module_extraction_contract_extraction_contract -- --nocapture` passes
-- [ ] `cargo test -p kamn-node --test main_module_extraction_contract -- --nocapture` passes
-- [ ] `python3 scripts/ci/check_touched_rust_size_policy.py --repo-root /tmp/kamn-6784-remote --base-ref origin/main --output-json /tmp/main-module-extraction-contract-size.json` returns `policy_decision=GO`
+- [x] `crates/kamn-node/tests/main_module_extraction_contract.rs` is reduced to a thin root shell at or below 180 LOC
+- [x] Root shell wires bounded sibling modules for the current main-module extraction concerns and shared support
+- [x] All extracted files touched by the split remain at or below 200 LOC
+- [x] `cargo test -p kamn-node --test main_module_extraction_contract_extraction_contract -- --nocapture` passes
+- [x] `cargo test -p kamn-node --test main_module_extraction_contract -- --nocapture` passes
+- [x] `python3 scripts/ci/check_touched_rust_size_policy.py --repo-root /tmp/kamn-6784-remote --base-ref origin/main --output-json /tmp/main-module-extraction-contract-size.json` returns `policy_decision=GO`
 
 ## Files to touch
 - `specs/6822-split-main-module-extraction-contract.md`
@@ -44,3 +44,12 @@ Split `crates/kamn-node/tests/main_module_extraction_contract.rs` into a thin ro
 3. Run the extraction contract target
 4. Run the real `main_module_extraction_contract` target
 5. Run the touched-Rust size checker against `origin/main`
+
+## Phase 6 evidence
+- `cargo test -p kamn-node --test main_module_extraction_contract_extraction_contract -- --nocapture`
+- `cargo test -p kamn-node --test main_module_extraction_contract -- --nocapture`
+- `python3 scripts/ci/check_touched_rust_size_policy.py --repo-root /tmp/kamn-6784-remote --base-ref origin/main --output-json /tmp/main-module-extraction-contract-size.json`
+- Result: `policy_decision=GO`
+
+## Deviations
+- None
